pyproject.toml
[project]
name = "dolores"
version = "0.1.11"
description = "Discord bot named Dolores"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apprise>=1.9.3",
    "hikari>=2.3.3",
    "hikari-lightbulb>=3.0.1",
    "hikari-miru>=4.2.0",
    "lavalink>=5.9.0",
    "openai>=1.75.0",
    "pydantic-ai>=0.0.29",
    "python-dotenv>=1.1.0",
    "requests==2.32.3",
    "tenacity>=9.0.0",
]
